feat: Pay and Invite to Pay docs (#2444)
