+++ api.planx.uk/modules/pay/service/inviteToPay/createPaymentSendEvents.test.ts
@@ -1,10 +1,10 @@
 import supertest from "supertest";
-import app from "../server";
-import { createScheduledEvent } from "../lib/hasura/metadata";
-import { queryMock } from "../tests/graphqlQueryMock";
-import { flowWithInviteToPay } from "../tests/mocks/inviteToPayData";
+import app from "../../../../server";
+import { createScheduledEvent } from "../../../../lib/hasura/metadata";
+import { queryMock } from "../../../../tests/graphqlQueryMock";
+import { flowWithInviteToPay } from "../../../../tests/mocks/inviteToPayData";
 
-jest.mock("../lib/hasura/metadata");
+jest.mock("../../../../lib/hasura/metadata");
 const mockedCreateScheduledEvent = createScheduledEvent as jest.MockedFunction<
   typeof createScheduledEvent
 >;
