+++ api.planx.uk/modules/pay/service/inviteToPay/createPaymentSendEvents.ts
@@ -1,13 +1,13 @@
 import { ComponentType } from "@opensystemslab/planx-core/types";
 import { NextFunction, Request, Response } from "express";
 import { gql } from "graphql-request";
-import { $api, $public } from "../client";
+import { $api, $public } from "../../../../client";
 import {
   ScheduledEventResponse,
   createScheduledEvent,
-} from "../lib/hasura/metadata";
-import { getMostRecentPublishedFlow } from "../helpers";
-import { Flow, Node, Team } from "../types";
+} from "../../../../lib/hasura/metadata";
+import { getMostRecentPublishedFlow } from "../../../../helpers";
+import { Flow, Node, Team } from "../../../../types";
 
 enum Destination {
   BOPS = "bops",
