+++ api.planx.uk/modules/pay/service/inviteToPay/index.ts
@@ -1,4 +1,3 @@
-export * from "./inviteToPay";
 export * from "./paymentRequest";
 export * from "./sendPaymentEmail";
 export * from "./sendConfirmationEmail";
